add pytest to project, move models to their own folder
