+++ pyproject.toml
@@ -9,6 +9,7 @@ python = "^3.8"
 Flask = "^2.2.5"
 python-dotenv = "^0.14.0"
 requests = "^2.31.0"
+pytest = "^8.1.1"
 
 [tool.poetry.dev-dependencies]
